+++ source/extensions/omni.isaac.core/docs/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [3.19.0] - 2024-07-26
+### Added
+- Added functions to find missing and incorrect semantic labels
+- Added function to tally semantic labels in a scene
+
 ## [3.18.1] - 2024-07-18
 ### Fixed
 - Fixed ArticulationController to handle joint indices along with Nan values in the target actions.